@@ -16,7 +16,7 @@ project_urls =
 
 [options]
 packages = find:
-python_requires = >=3.5
+python_requires = >=3.6
 py_modules =
     find_engine
 include_package_data = True
